+++ src/View/Helper/FormHelper.php
@@ -45,11 +45,7 @@ class FormHelper extends BaseFormHelper
     protected bool $validation = true;
 
     /**
-     * Construct the widgets and binds the default context providers.
-     *
-     * This method only rewrites the default config.
-     * @param \Cake\View\View $view The View this helper is being attached to
-     * @param array<string, mixed> $config Configuration settings for the helper
+     * @inheritDoc
      */
     public function __construct(View $view, array $config = [])
     {
@@ -87,11 +83,7 @@ class FormHelper extends BaseFormHelper
     }
 
     /**
-     * Returns the input type that was guessed for the provided fieldName, based on the internal type it is associated
-     *  too, its name and the variables that can be found in the view template
-     * @param string $fieldName the name of the field to guess a type for
-     * @param array<string, mixed> $options the options passed to the input method
-     * @return string
+     * @inheritDoc
      */
     protected function _inputType(string $fieldName, array $options): string
     {
@@ -102,10 +94,7 @@ class FormHelper extends BaseFormHelper
     }
 
     /**
-     * Generate label for input
-     * @param string $fieldName The name of the field to generate label for
-     * @param array<string, mixed> $options Options list
-     * @return string|false Generated label element or false
+     * @inheritDoc
      */
     protected function _getLabel(string $fieldName, array $options)
     {
@@ -305,10 +294,7 @@ class FormHelper extends BaseFormHelper
     }
 
     /**
-     * Closes an HTML form, cleans up values set by FormHelper::create(), and writes hidden input fields where appropriate
-     * @param array<string, mixed> $secureAttributes Secure attributes which will be passed as HTML attributes into the
-     *  hidden input elements generated for the Security Component
-     * @return string A closing FORM tag
+     * @inheritDoc
      */
     public function end(array $secureAttributes = []): string
     {
@@ -335,13 +321,7 @@ class FormHelper extends BaseFormHelper
     }
 
     /**
-     * Returns a formatted SELECT element
-     * @param string $fieldName Name attribute of the SELECT
-     * @param iterable $options Array of the OPTION elements (as 'value'=>'Text' pairs) to be used in the SELECT element
-     * @param array<string, mixed> $attributes The HTML attributes of the select element
-     * @return string Formatted SELECT element
-     * @link https://book.cakephp.org/4/en/views/helpers/form.html#creating-select-pickers
-     * @see \Cake\View\Helper\FormHelper::select() for all available options
+     * @inheritDoc
      */
     public function select(string $fieldName, iterable $options = [], array $attributes = []): string
     {